refactor: fixes #39 (separating transforms by their ontology)
