+++ pytoda/smiles/transforms.py
@@ -1,7 +1,6 @@
 """SMILES transforms."""
 import logging
 import re
-from copy import deepcopy
 
 import numpy as np
 import torch
@@ -15,7 +14,7 @@ from ..transforms import Transform
 from ..types import Indexes
 from .smiles_language import SMILESLanguage
 
-logger = logging.getLogger('SMILES_transforms')
+logger = logging.getLogger('pytoda_SMILES_transforms')
 
 
 class SMILESToTokenIndexes(Transform):
@@ -41,73 +40,6 @@ class SMILESToTokenIndexes(Transform):
             Indexes: indexes representation for the SMILES provided.
         """
         return self.smiles_language.smiles_to_token_indexes(smiles)
-
-
-class LeftPadding(Transform):
-    """Left pad token indexes."""
-
-    def __init__(self, padding_length: int, padding_index: int) -> None:
-        """
-        Initialize a left padding token indexes object.
-
-        Args:
-            padding_length (int): length of the padding.
-            padding_index (int): padding index.
-        """
-        self.padding_length = padding_length
-        self.padding_index = padding_index
-
-    def __call__(self, token_indexes: Indexes) -> Indexes:
-        """
-        Apply the transform.
-
-        Args:
-            token_indexes (Indexes): token indexes.
-
-        Returns:
-            Indexes: left padded indexes representation.
-        """
-        if self.padding_length < len(token_indexes):
-            logger.warning(
-                f'\n{token_indexes} is longer than padding length '
-                f'({self.padding_length}). End of string will be stripped off.'
-            )
-            return token_indexes[:self.padding_length]
-        else:
-            return (
-                (self.padding_length - len(token_indexes)) *
-                [self.padding_index] + token_indexes
-            )
-
-
-class ToTensor(Transform):
-    """Transform token indexes to torch tensor."""
-
-    def __init__(self, device, dtype=torch.short) -> None:
-        """
-        Initialize a token indexes to tensor object.
-
-        Args:
-            dtype (torch.dtype): data type. Defaults to torch.short.
-            device (torch.device): device where the tensors are stored.
-            Defaults to gpu, if available.
-        """
-        self.dtype = torch.short
-        self.device = device
-
-    def __call__(self, token_indexes: Indexes) -> torch.Tensor:
-        """
-        Apply the transform.
-
-        Args:
-            token_indexes (Indexes): token indexes.
-
-        Returns:
-            torch.Tensor: tensor representation of the token indexes.
-        """
-        return torch.tensor(
-            token_indexes, dtype=self.dtype, device=self.device
-        ).view(-1, 1).squeeze()
 
 
 class RemoveIsomery(Transform):
@@ -471,27 +403,6 @@ class AugmentTensor(Transform):
 
         augmented = torch.cat(augmented, dim=0)
         return augmented
-
-
-class Randomize(Transform):
-    """Randomize a molecule by truly shuffling all tokens."""
-
-    def __call__(self, tokens: Indexes) -> Indexes:
-        """
-        Intialize SMILES randomizer.
-
-        NOTE: Must not apply this transformation on SMILES string, only on the
-            tokenized, numerical vectors (i.e. after SMILESToTokenIndexes)
-
-        Args:
-            tokens (Indexes): indexes representation for the SMILES to be
-                randomized.
-        Returns:
-           Indexes: shuffled indexes representation of the molecule
-        """
-        smiles_tokens = deepcopy(tokens)
-        np.random.shuffle(smiles_tokens)
-        return smiles_tokens
 
 
 class Selfies(Transform):
