+++ pytoda/transforms.py
@@ -1,10 +1,20 @@
 """Transform utilities."""
+import logging
+import random
+from copy import deepcopy
 from typing import Any
-from .types import TransformList
+
+import numpy as np
+import torch
+
+from .types import Indexes, TransformList
+
+logger = logging.getLogger('pytoda_transforms')
 
 
 class Transform(object):
     """Transform abstract class."""
+
     def __call__(self, sample: Any) -> Any:
         """
         Apply the transformation.
@@ -16,6 +26,104 @@ class Transform(object):
             Any: the transformed sample.
         """
         raise NotImplementedError
+
+
+class LeftPadding(Transform):
+    """Left pad token indexes."""
+
+    def __init__(self, padding_length: int, padding_index: int) -> None:
+        """
+        Initialize a left padding token indexes object.
+
+        Args:
+            padding_length (int): length of the padding.
+            padding_index (int): padding index.
+        """
+        self.padding_length = padding_length
+        self.padding_index = padding_index
+
+    def __call__(self, token_indexes: Indexes) -> Indexes:
+        """
+        Apply the transform.
+
+        Args:
+            token_indexes (Indexes): token indexes.
+
+        Returns:
+            Indexes: left padded indexes representation.
+        """
+
+        return (
+            (self.padding_length - len(token_indexes)) * [self.padding_index] +
+            token_indexes
+        )
+
+
+class ToTensor(Transform):
+    """Transform token indexes to torch tensor."""
+
+    def __init__(self, device, dtype=torch.short) -> None:
+        """
+        Initialize a token indexes to tensor object.
+
+        Args:
+            dtype (torch.dtype): data type. Defaults to torch.short.
+            device (torch.device): device where the tensors are stored.
+            Defaults to gpu, if available.
+        """
+        self.dtype = torch.short
+        self.device = device
+
+    def __call__(self, token_indexes: Indexes) -> torch.Tensor:
+        """
+        Apply the transform.
+
+        Args:
+            token_indexes (Indexes): token indexes.
+
+        Returns:
+            torch.Tensor: tensor representation of the token indexes.
+        """
+        return torch.tensor(
+            token_indexes, dtype=self.dtype, device=self.device
+        ).view(-1, 1).squeeze()
+
+
+class Randomize(Transform):
+    """Randomize a molecule by truly shuffling all tokens."""
+
+    def __call__(self, tokens: Indexes) -> Indexes:
+        """
+        Intialize SMILES randomizer.
+
+        NOTE: Must not apply this transformation on SMILES string, only on the
+            tokenized, numerical vectors (i.e. after SMILESToTokenIndexes)
+
+        Args:
+            tokens (Indexes): indexes representation for the SMILES to be
+                randomized.
+        Returns:
+           Indexes: shuffled indexes representation of the molecule
+        """
+        smiles_tokens = deepcopy(tokens)
+        np.random.shuffle(smiles_tokens)
+        return smiles_tokens
+
+
+class AugmentByReversing(Transform):
+    """Augment an amino acid sequence by (eventually) flipping order"""
+
+    def __call__(self, sequence: str) -> str:
+        """
+        Apply the transform.
+
+        Args:
+            sequnce (str): a sequence representation.
+
+        Returns:
+            str: Either the sequence itself, or the revesed sequence.
+        """
+        return sequence[::-1] if round(random.random()) else sequence
 
 
 class Compose(Transform):
